# src/backend/openapi_debug.json
{"openapi":"3.1.0","info":{"title":"Lambda.hu API","description":"API for the Lambda.hu building material intelligence system.","version":"1.0.0"},"paths":{"/api/v1/products":{"get":{"summary":"Read Products","operationId":"read_products_api_v1_products_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"},"title":"Response Read Products Api V1 Products Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/categories":{"get":{"summary":"Read Categories","operationId":"read_categories_api_v1_categories_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Category"},"title":"Response Read Categories Api V1 Categories Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"Redirects root to the main search interface.","operationId":"root__get","responses":{"307":{"description":"Successful Response"}}}},"/search/rag":{"post":{"summary":"Perform a RAG search","description":"Végrehajt egy szemantikus keresést a vektor adatbázisban (ChromaDB)\na felhasználói bevitelen alapulva.\n\nEz a végpont a fő kereső funkciót szolgálja ki.","operationId":"rag_search_search_rag_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Category":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"parent_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Id"},"id":{"type":"integer","title":"Id"},"children":{"items":{"$ref":"#/components/schemas/Category"},"type":"array","title":"Children","default":[]}},"type":"object","required":["name","id"],"title":"Category"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Manufacturer":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"website":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Website"},"country":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Country"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["name","id"],"title":"Manufacturer"},"Product":{"properties":{"name":{"type":"string","title":"Name"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Price"},"technical_specs":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Technical Specs"},"raw_specs":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Raw Specs"},"full_text_content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Full Text Content"},"category_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Category Id"},"manufacturer_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Manufacturer Id"},"id":{"type":"integer","title":"Id"},"manufacturer":{"anyOf":[{"$ref":"#/components/schemas/Manufacturer"},{"type":"null"}]},"category":{"anyOf":[{"$ref":"#/components/schemas/Category"},{"type":"null"}]},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Updated At"}},"type":"object","required":["name","id","created_at"],"title":"Product"},"SearchRequest":{"properties":{"query":{"type":"string","title":"Query"},"limit":{"type":"integer","title":"Limit","default":10}},"type":"object","required":["query"],"title":"SearchRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}